requirements.txt
absl-py>=0.15.0
astor>=0.8.1
cached-property>=1.5.2
cachetools>=5.0.0
certifi>=2022.6.15
chardet>=4.0.0
click>=8.0.3
colorama>=0.4.5
dataclasses; python_version<'3.7'
Flask>=2.0.0
gast>=0.5.3
google-auth>=2.6.0
google-auth-oauthlib>=0.5.0
google-pasta>=0.2.0
grpcio>=1.46.3
h5py>=3.7.0
idna>=3.3
importlib-metadata>=4.8.1; python_version<"3.8"
itsdangerous>=2.1.0
Jinja2>=3.1.2
joblib>=1.1.0
Keras-Applications>=1.0.8
Keras-Preprocessing>=1.1.2
Markdown>=3.4.1
MarkupSafe>=2.1.1
numpy>=1.22.0
oauthlib>=3.2.0
opt-einsum>=3.3.0
pickle-mixin>=1.0.2
Pillow>=9.2.0
protobuf>=3.19.6
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
requests>=2.27.1
requests-oauthlib>=1.3.1
rsa>=4.8
scikit-learn>=1.0.2
scipy>=1.7.3
six>=1.16.0
tensorboard>=2.9.0
tensorflow>=2.9.1
tensorflow-estimator>=2.9.0
termcolor>=1.1.0
typing-extensions>=4.1.1
urllib3>=1.26.9
Werkzeug>=2.1.2
wrapt>=1.14.1
zipp>=3.8.1
pandas
matplotlib
opencv-python
reportlab
